Adds check for gcompat lib on alpine for oc install (#26)

+++ scripts/setup-oc.sh
@@ -5,10 +5,6 @@ SCRIPT_DIR=$(cd $(dirname "$0"); pwd -P)
 DEST_DIR="$1"
 TYPE="$2"
 
-RELEASE=$(curl -s "https://api.github.com/repos/bitnami-labs/sealed-secrets/releases/latest" | "${DEST_DIR}/jq" -r '.tag_name')
-
-SHORT_RELEASE=$(echo "${RELEASE}" | sed -E "s/v//g")
-
 FILETYPE="linux"
 if [[ "${TYPE}" == "macos" ]]; then
   FILETYPE="mac"
@@ -17,13 +13,13 @@ fi
 URL="https://mirror.openshift.com/pub/openshift-v4/x86_64/clients/ocp/stable/openshift-client-${FILETYPE}.tar.gz"
 
 CMD_NAME="oc"
-if [[ "${TYPE}" == "alpine" ]]; then
+if [[ "${TYPE}" == "alpine" ]] && [[ ! -f /lib/libgcompat.so.0 ]]; then
   CMD_NAME="oc-bin"
 fi
 
 "${SCRIPT_DIR}/setup-binary-from-tgz.sh" "${DEST_DIR}" "${CMD_NAME}" "${URL}" oc
 
-if [[ "${TYPE}" == "alpine" ]]; then
+if [[ "${TYPE}" == "alpine" ]] && [[ ! -f /lib/libgcompat.so.0 ]]; then
   echo "/lib/ld-musl-x86_64.so.1 --library-path /lib ${DEST_DIR}/oc-bin \$@" > ${DEST_DIR}/oc
   chmod +x "${DEST_DIR}/oc"
 fi
